requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
certifi==2024.7.4
colorlog==6.8.2
fluent==0.10.0
fluent.runtime==0.4.0
fluent.syntax==0.19.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
SQLAlchemy==2.0.31
stubs==1.0.0
tomli==2.0.1
typing_extensions==4.12.2
yarl==1.9.4
